Clean up code, bring into alignment with style/vocab
