+++ js/daily-call-container.js
@@ -12,12 +12,12 @@ template.innerHTML = `
   `;
 
 class DailyCallContainer extends HTMLElement {
-    constructor(options = {}) {
-        super();
-        const shadow = this.attachShadow({ mode: 'open' });
-        this.shadowRoot.appendChild(template.content.cloneNode(true))
-    }
-    connectedCallback() {}
+  constructor(options = {}) {
+    super();
+    const shadow = this.attachShadow({ mode: 'open' });
+    this.shadowRoot.appendChild(template.content.cloneNode(true))
+  }
+  connectedCallback() {}
 }
 
 window.customElements.define('daily-call', DailyCallContainer);
